refactor prase NodeStore into separate file
