+++ src/check/parse/Parser.zig
@@ -1,16 +1,17 @@
 const std = @import("std");
+const tracy = @import("../../tracy.zig");
+const tokenize = @import("tokenize.zig");
+const collections = @import("../../collections.zig");
 
 const AST = @import("AST.zig");
 const Node = @import("Node.zig");
+const NodeStore = @import("NodeStore.zig");
 const NodeList = AST.NodeList;
-
-const tracy = @import("../../tracy.zig");
-const tokenize = @import("tokenize.zig");
 const TokenizedBuffer = tokenize.TokenizedBuffer;
 const Token = tokenize.Token;
 const TokenIdx = Token.Idx;
 
-const exitOnOom = @import("../../collections/utils.zig").exitOnOom;
+const exitOnOom = collections.utils.exitOnOom;
 
 /// A parser which tokenizes and parses source code into an abstract syntax tree.
 pub const Parser = @This();
@@ -18,14 +19,14 @@ pub const Parser = @This();
 gpa: std.mem.Allocator,
 pos: TokenIdx,
 tok_buf: TokenizedBuffer,
-store: AST.NodeStore,
+store: NodeStore,
 scratch_nodes: std.ArrayListUnmanaged(Node.Idx),
 diagnostics: std.ArrayListUnmanaged(AST.Diagnostic),
 
 /// init the parser from a buffer of tokens
 pub fn init(tokens: TokenizedBuffer) Parser {
     const estimated_node_count = tokens.tokens.len;
-    const store = AST.NodeStore.initCapacity(tokens.env.gpa, estimated_node_count);
+    const store = NodeStore.initCapacity(tokens.env.gpa, estimated_node_count);
 
     return Parser{
         .gpa = tokens.env.gpa,
@@ -158,7 +159,7 @@ pub fn parseFile(self: *Parser) void {
 
     self.store.emptyScratch();
     _ = self.store.addFile(.{
-        .header = AST.NodeStore.HeaderIdx{ .id = 0 },
+        .header = NodeStore.HeaderIdx{ .id = 0 },
         .statements = .{ .span = .{
             .start = 0,
             .len = 0,
@@ -194,7 +195,7 @@ pub fn parseFile(self: *Parser) void {
 /// Parses the items of type T until we encounter end_token, with each item separated by a Comma token
 ///
 /// Returns the ending position of the collection
-fn parseCollectionSpan(self: *Parser, comptime T: type, end_token: Token.Tag, scratch_fn: fn (*AST.NodeStore, T) void, parser: fn (*Parser) T) ExpectError!u32 {
+fn parseCollectionSpan(self: *Parser, comptime T: type, end_token: Token.Tag, scratch_fn: fn (*NodeStore, T) void, parser: fn (*Parser) T) ExpectError!u32 {
     while (self.peek() != end_token and self.peek() != .EndOfFile) {
         scratch_fn(&self.store, parser(self));
         self.expect(.Comma) catch {
@@ -231,7 +232,7 @@ fn debugToken(self: *Parser, window: usize) void {
 /// provides_entry :: [LowerIdent|UpperIdent] Comma Newline*
 /// package_entry :: LowerIdent Comma "platform"? String Comma
 /// app_header :: KwApp Newline* OpenSquare provides_entry* CloseSquare OpenCurly package_entry CloseCurly
-pub fn parseHeader(self: *Parser) AST.NodeStore.HeaderIdx {
+pub fn parseHeader(self: *Parser) NodeStore.HeaderIdx {
     switch (self.peek()) {
         .KwApp => {
             return self.parseAppHeader();
@@ -250,7 +251,7 @@ pub fn parseHeader(self: *Parser) AST.NodeStore.HeaderIdx {
             return self.parsePlatformHeader();
         },
         else => {
-            return self.pushMalformed(AST.NodeStore.HeaderIdx, .missing_header, self.pos);
+            return self.pushMalformed(NodeStore.HeaderIdx, .missing_header, self.pos);
         },
     }
 }
@@ -265,7 +266,7 @@ pub fn parseHeader(self: *Parser) AST.NodeStore.HeaderIdx {
 ///     packages { foo: "../foo.roc" }
 ///     imports []
 ///     provides [main_for_host]
-pub fn parsePlatformHeader(self: *Parser) AST.NodeStore.HeaderIdx {
+pub fn parsePlatformHeader(self: *Parser) NodeStore.HeaderIdx {
     const start = self.pos;
     std.debug.assert(self.peek() == .KwPlatform);
     self.advance(); // Advance past KwPlatform
@@ -273,7 +274,7 @@ pub fn parsePlatformHeader(self: *Parser) AST.NodeStore.HeaderIdx {
     // Get name
     self.expect(.StringStart) catch {
         return self.pushMalformed(
-            AST.NodeStore.HeaderIdx,
+            NodeStore.HeaderIdx,
             .expected_platform_name_start,
             self.pos,
         );
@@ -281,14 +282,14 @@ pub fn parsePlatformHeader(self: *Parser) AST.NodeStore.HeaderIdx {
     const name = self.pos;
     self.expect(.StringPart) catch {
         return self.pushMalformed(
-            AST.NodeStore.HeaderIdx,
+            NodeStore.HeaderIdx,
             .expected_platform_name_string,
             self.pos,
         );
     };
     self.expect(.StringEnd) catch {
         return self.pushMalformed(
-            AST.NodeStore.HeaderIdx,
+            NodeStore.HeaderIdx,
             .expected_platform_name_end,
             self.pos,
         );
@@ -296,7 +297,7 @@ pub fn parsePlatformHeader(self: *Parser) AST.NodeStore.HeaderIdx {
 
     self.expect(.KwRequires) catch {
         return self.pushMalformed(
-            AST.NodeStore.HeaderIdx,
+            NodeStore.HeaderIdx,
             .expected_requires,
             self.pos,
         );
@@ -305,21 +306,21 @@ pub fn parsePlatformHeader(self: *Parser) AST.NodeStore.HeaderIdx {
     const rigids_start = self.pos;
     self.expect(.OpenCurly) catch {
         return self.pushMalformed(
-            AST.NodeStore.HeaderIdx,
+            NodeStore.HeaderIdx,
             .expected_requires_rigids_open_curly,
             self.pos,
         );
     };
     const rigids_top = self.store.scratchExposedItemTop();
     const rigids_end = self.parseCollectionSpan(
-        AST.NodeStore.ExposedItemIdx,
+        NodeStore.ExposedItemIdx,
         .CloseCurly,
-        AST.NodeStore.addScratchExposedItem,
+        NodeStore.addScratchExposedItem,
         Parser.parseExposedItem,
     ) catch {
         self.store.clearScratchExposedItemsFrom(rigids_start);
         return self.pushMalformed(
-            AST.NodeStore.HeaderIdx,
+            NodeStore.HeaderIdx,
             .expected_requires_rigids_close_curly,
             rigids_start,
         );
@@ -340,20 +341,20 @@ pub fn parsePlatformHeader(self: *Parser) AST.NodeStore.HeaderIdx {
     const signatures_start = self.pos;
     self.expect(.OpenCurly) catch {
         return self.pushMalformed(
-            AST.NodeStore.HeaderIdx,
+            NodeStore.HeaderIdx,
             .expected_requires_signatures_open_curly,
             self.pos,
         );
     };
     const signatures_top = self.store.scratchAnnoRecordFieldTop();
     const signatures_end = self.parseCollectionSpan(
-        AST.NodeStore.AnnoRecordFieldIdx,
+        NodeStore.AnnoRecordFieldIdx,
         .CloseCurly,
-        AST.NodeStore.addScratchAnnoRecordField,
+        NodeStore.addScratchAnnoRecordField,
         Parser.parseAnnoRecordField,
     ) catch {
         return self.pushMalformed(
-            AST.NodeStore.HeaderIdx,
+            NodeStore.HeaderIdx,
             .expected_requires_signatures_close_curly,
             rigids_start,
         );
@@ -370,7 +371,7 @@ pub fn parsePlatformHeader(self: *Parser) AST.NodeStore.HeaderIdx {
     // Get exposes
     self.expect(.KwExposes) catch {
         return self.pushMalformed(
-            AST.NodeStore.HeaderIdx,
+            NodeStore.HeaderIdx,
             .expected_exposes,
             self.pos,
         );
@@ -378,21 +379,21 @@ pub fn parsePlatformHeader(self: *Parser) AST.NodeStore.HeaderIdx {
     const exposes_start = self.pos;
     self.expect(.OpenSquare) catch {
         return self.pushMalformed(
-            AST.NodeStore.HeaderIdx,
+            NodeStore.HeaderIdx,
             .expected_exposes_open_square,
             self.pos,
         );
     };
     const exposes_top = self.store.scratchExposedItemTop();
     const exposes_end = self.parseCollectionSpan(
-        AST.NodeStore.ExposedItemIdx,
+        NodeStore.ExposedItemIdx,
         .CloseSquare,
-        AST.NodeStore.addScratchExposedItem,
+        NodeStore.addScratchExposedItem,
         Parser.parseExposedItem,
     ) catch {
         self.store.clearScratchExposedItemsFrom(exposes_start);
         return self.pushMalformed(
-            AST.NodeStore.HeaderIdx,
+            NodeStore.HeaderIdx,
             .expected_exposes_close_square,
             exposes_start,
         );
@@ -409,7 +410,7 @@ pub fn parsePlatformHeader(self: *Parser) AST.NodeStore.HeaderIdx {
     // Get packages
     self.expect(.KwPackages) catch {
         return self.pushMalformed(
-            AST.NodeStore.HeaderIdx,
+            NodeStore.HeaderIdx,
             .expected_imports,
             self.pos,
         );
@@ -418,20 +419,20 @@ pub fn parsePlatformHeader(self: *Parser) AST.NodeStore.HeaderIdx {
     const packages_top = self.store.scratchRecordFieldTop();
     self.expect(.OpenCurly) catch {
         return self.pushMalformed(
-            AST.NodeStore.HeaderIdx,
+            NodeStore.HeaderIdx,
             .expected_packages_open_curly,
             self.pos,
         );
     };
     const packages_end = self.parseCollectionSpan(
-        AST.NodeStore.RecordFieldIdx,
+        NodeStore.RecordFieldIdx,
         .CloseCurly,
-        AST.NodeStore.addScratchRecordField,
+        NodeStore.addScratchRecordField,
         Parser.parseRecordField,
     ) catch {
         self.store.clearScratchRecordFieldsFrom(packages_top);
         return self.pushMalformed(
-            AST.NodeStore.HeaderIdx,
+            NodeStore.HeaderIdx,
             .expected_packages_close_curly,
             self.pos,
         );
@@ -448,7 +449,7 @@ pub fn parsePlatformHeader(self: *Parser) AST.NodeStore.HeaderIdx {
     // Get provides
     self.expect(.KwProvides) catch {
         return self.pushMalformed(
-            AST.NodeStore.HeaderIdx,
+            NodeStore.HeaderIdx,
             .expected_provides,
             self.pos,
         );
@@ -456,21 +457,21 @@ pub fn parsePlatformHeader(self: *Parser) AST.NodeStore.HeaderIdx {
     const provides_start = self.pos;
     self.expect(.OpenSquare) catch {
         return self.pushMalformed(
-            AST.NodeStore.HeaderIdx,
+            NodeStore.HeaderIdx,
             .expected_provides_open_square,
             self.pos,
         );
     };
     const provides_top = self.store.scratchExposedItemTop();
     const end = self.parseCollectionSpan(
-        AST.NodeStore.ExposedItemIdx,
+        NodeStore.ExposedItemIdx,
         .CloseSquare,
-        AST.NodeStore.addScratchExposedItem,
+        NodeStore.addScratchExposedItem,
         Parser.parseExposedItem,
     ) catch {
         self.store.clearScratchExposedItemsFrom(provides_start);
         return self.pushMalformed(
-            AST.NodeStore.HeaderIdx,
+            NodeStore.HeaderIdx,
             .expected_provides_close_square,
             provides_start,
         );
@@ -498,7 +499,7 @@ pub fn parsePlatformHeader(self: *Parser) AST.NodeStore.HeaderIdx {
 /// parse an `.roc` package header
 ///
 /// e.g. `package [ foo ] { something: "package/path/main.roc" }`
-pub fn parsePackageHeader(self: *Parser) AST.NodeStore.HeaderIdx {
+pub fn parsePackageHeader(self: *Parser) NodeStore.HeaderIdx {
     const start = self.pos;
 
     std.debug.assert(self.peek() == .KwPackage);
@@ -507,18 +508,18 @@ pub fn parsePackageHeader(self: *Parser) AST.NodeStore.HeaderIdx {
     // Get Exposes
     const exposes_start = self.pos;
     self.expect(.OpenSquare) catch {
-        return self.pushMalformed(AST.NodeStore.HeaderIdx, .expected_provides_open_square, start);
+        return self.pushMalformed(NodeStore.HeaderIdx, .expected_provides_open_square, start);
     };
     const scratch_top = self.store.scratchExposedItemTop();
-    const exposes_end = self.parseCollectionSpan(AST.NodeStore.ExposedItemIdx, .CloseSquare, AST.NodeStore.addScratchExposedItem, Parser.parseExposedItem) catch {
+    const exposes_end = self.parseCollectionSpan(NodeStore.ExposedItemIdx, .CloseSquare, NodeStore.addScratchExposedItem, Parser.parseExposedItem) catch {
         while (self.peek() != .CloseSquare and self.peek() != .EndOfFile) {
             self.advance();
         }
         self.expect(.CloseSquare) catch {
-            return self.pushMalformed(AST.NodeStore.HeaderIdx, .header_expected_close_square, start);
+            return self.pushMalformed(NodeStore.HeaderIdx, .header_expected_close_square, start);
         };
         self.store.clearScratchExposedItemsFrom(scratch_top);
-        return self.pushMalformed(AST.NodeStore.HeaderIdx, .import_exposing_no_close, start);
+        return self.pushMalformed(NodeStore.HeaderIdx, .import_exposing_no_close, start);
     };
     const exposes_span = self.store.exposedItemSpanFrom(scratch_top);
     const exposes = self.store.addCollection(.collection_exposed, .{
@@ -532,12 +533,12 @@ pub fn parsePackageHeader(self: *Parser) AST.NodeStore.HeaderIdx {
     // Get Packages
     const packages_start = self.pos;
     self.expect(.OpenCurly) catch {
-        return self.pushMalformed(AST.NodeStore.HeaderIdx, .expected_package_platform_open_curly, start);
+        return self.pushMalformed(NodeStore.HeaderIdx, .expected_package_platform_open_curly, start);
     };
     const fields_scratch_top = self.store.scratchRecordFieldTop();
-    const end = self.parseCollectionSpan(AST.NodeStore.RecordFieldIdx, .CloseCurly, AST.NodeStore.addScratchRecordField, Parser.parseRecordField) catch {
+    const end = self.parseCollectionSpan(NodeStore.RecordFieldIdx, .CloseCurly, NodeStore.addScratchRecordField, Parser.parseRecordField) catch {
         self.store.clearScratchRecordFieldsFrom(fields_scratch_top);
-        return self.pushMalformed(AST.NodeStore.HeaderIdx, .expected_package_platform_close_curly, start);
+        return self.pushMalformed(NodeStore.HeaderIdx, .expected_package_platform_close_curly, start);
     };
     const packages_span = self.store.recordFieldSpanFrom(fields_scratch_top);
     const packages = self.store.addCollection(.collection_packages, .{
@@ -550,7 +551,7 @@ pub fn parsePackageHeader(self: *Parser) AST.NodeStore.HeaderIdx {
 
     self.advance();
 
-    const header = AST.NodeStore.Header{ .package = .{
+    const header = NodeStore.Header{ .package = .{
         .exposes = exposes,
         .packages = packages,
         .region = .{ .start = start, .end = end },
@@ -562,7 +563,7 @@ pub fn parsePackageHeader(self: *Parser) AST.NodeStore.HeaderIdx {
 /// Parse a Roc Hosted header
 ///
 /// e.g. `hosted [foo]`
-fn parseHostedHeader(self: *Parser) AST.NodeStore.HeaderIdx {
+fn parseHostedHeader(self: *Parser) NodeStore.HeaderIdx {
     std.debug.assert(self.peek() == .KwHosted);
 
     const start = self.pos;
@@ -572,18 +573,18 @@ fn parseHostedHeader(self: *Parser) AST.NodeStore.HeaderIdx {
     // Get exposes
     const exposes_start = self.pos;
     self.expect(.OpenSquare) catch {
-        return self.pushMalformed(AST.NodeStore.HeaderIdx, .header_expected_open_square, self.pos);
+        return self.pushMalformed(NodeStore.HeaderIdx, .header_expected_open_square, self.pos);
     };
     const scratch_top = self.store.scratchExposedItemTop();
-    const end = self.parseCollectionSpan(AST.NodeStore.ExposedItemIdx, .CloseSquare, AST.NodeStore.addScratchExposedItem, Parser.parseExposedItem) catch {
+    const end = self.parseCollectionSpan(NodeStore.ExposedItemIdx, .CloseSquare, NodeStore.addScratchExposedItem, Parser.parseExposedItem) catch {
         while (self.peek() != .CloseSquare and self.peek() != .EndOfFile) {
             self.advance();
         }
         self.expect(.CloseSquare) catch {
-            return self.pushMalformed(AST.NodeStore.HeaderIdx, .header_expected_close_square, self.pos);
+            return self.pushMalformed(NodeStore.HeaderIdx, .header_expected_close_square, self.pos);
         };
         self.store.clearScratchExposedItemsFrom(scratch_top);
-        return self.pushMalformed(AST.NodeStore.HeaderIdx, .import_exposing_no_close, self.pos);
+        return self.pushMalformed(NodeStore.HeaderIdx, .import_exposing_no_close, self.pos);
     };
     const exposes_span = self.store.exposedItemSpanFrom(scratch_top);
     const exposes = self.store.addCollection(.collection_exposed, .{
@@ -603,7 +604,7 @@ fn parseHostedHeader(self: *Parser) AST.NodeStore.HeaderIdx {
 /// parse a Roc module header
 ///
 /// e.g. `module [foo]`
-fn parseModuleHeader(self: *Parser) AST.NodeStore.HeaderIdx {
+fn parseModuleHeader(self: *Parser) NodeStore.HeaderIdx {
     std.debug.assert(self.peek() == .KwModule);
 
     const start = self.pos;
@@ -613,18 +614,18 @@ fn parseModuleHeader(self: *Parser) AST.NodeStore.HeaderIdx {
     // Get exposes
     const exposes_start = self.pos;
     self.expect(.OpenSquare) catch {
-        return self.pushMalformed(AST.NodeStore.HeaderIdx, .header_expected_open_square, self.pos);
+        return self.pushMalformed(NodeStore.HeaderIdx, .header_expected_open_square, self.pos);
     };
     const scratch_top = self.store.scratchExposedItemTop();
-    const end = self.parseCollectionSpan(AST.NodeStore.ExposedItemIdx, .CloseSquare, AST.NodeStore.addScratchExposedItem, Parser.parseExposedItem) catch {
+    const end = self.parseCollectionSpan(NodeStore.ExposedItemIdx, .CloseSquare, NodeStore.addScratchExposedItem, Parser.parseExposedItem) catch {
         while (self.peek() != .CloseSquare and self.peek() != .EndOfFile) {
             self.advance();
         }
         self.expect(.CloseSquare) catch {
-            return self.pushMalformed(AST.NodeStore.HeaderIdx, .header_expected_close_square, self.pos);
+            return self.pushMalformed(NodeStore.HeaderIdx, .header_expected_close_square, self.pos);
         };
         self.store.clearScratchExposedItemsFrom(scratch_top);
-        return self.pushMalformed(AST.NodeStore.HeaderIdx, .import_exposing_no_close, self.pos);
+        return self.pushMalformed(NodeStore.HeaderIdx, .import_exposing_no_close, self.pos);
     };
     const exposes_span = self.store.exposedItemSpanFrom(scratch_top);
     const exposes = self.store.addCollection(.collection_exposed, .{
@@ -644,8 +645,8 @@ fn parseModuleHeader(self: *Parser) AST.NodeStore.HeaderIdx {
 /// parse an `.roc` application header
 ///
 /// e.g. `app [main!] { pf: "../some-platform.roc" }`
-pub fn parseAppHeader(self: *Parser) AST.NodeStore.HeaderIdx {
-    var platform: ?AST.NodeStore.RecordFieldIdx = null;
+pub fn parseAppHeader(self: *Parser) NodeStore.HeaderIdx {
+    var platform: ?NodeStore.RecordFieldIdx = null;
     const start = self.pos;
 
     std.debug.assert(self.peek() == .KwApp);
@@ -653,19 +654,19 @@ pub fn parseAppHeader(self: *Parser) AST.NodeStore.HeaderIdx {
 
     // Get provides
     self.expect(.OpenSquare) catch {
-        return self.pushMalformed(AST.NodeStore.HeaderIdx, .expected_provides_open_square, start);
+        return self.pushMalformed(NodeStore.HeaderIdx, .expected_provides_open_square, start);
     };
     const provides_start = self.pos;
     const scratch_top = self.store.scratchExposedItemTop();
-    const provides_end = self.parseCollectionSpan(AST.NodeStore.ExposedItemIdx, .CloseSquare, AST.NodeStore.addScratchExposedItem, Parser.parseExposedItem) catch {
+    const provides_end = self.parseCollectionSpan(NodeStore.ExposedItemIdx, .CloseSquare, NodeStore.addScratchExposedItem, Parser.parseExposedItem) catch {
         while (self.peek() != .CloseSquare and self.peek() != .EndOfFile) {
             self.advance();
         }
         self.expect(.CloseSquare) catch {
-            return self.pushMalformed(AST.NodeStore.HeaderIdx, .header_expected_close_square, start);
+            return self.pushMalformed(NodeStore.HeaderIdx, .header_expected_close_square, start);
         };
         self.store.clearScratchExposedItemsFrom(scratch_top);
-        return self.pushMalformed(AST.NodeStore.HeaderIdx, .import_exposing_no_close, start);
+        return self.pushMalformed(NodeStore.HeaderIdx, .import_exposing_no_close, start);
     };
     const provides_span = self.store.exposedItemSpanFrom(scratch_top);
     const provides_region = AST.Region{ .start = provides_start, .end = provides_end };
@@ -675,7 +676,7 @@ pub fn parseAppHeader(self: *Parser) AST.NodeStore.HeaderIdx {
     const fields_scratch_top = self.store.scratchRecordFieldTop();
     const packages_start = self.pos;
     self.expect(.OpenCurly) catch {
-        return self.pushMalformed(AST.NodeStore.HeaderIdx, .expected_package_platform_open_curly, start);
+        return self.pushMalformed(NodeStore.HeaderIdx, .expected_package_platform_open_curly, start);
     };
     var i: usize = 0;
 
@@ -683,24 +684,24 @@ pub fn parseAppHeader(self: *Parser) AST.NodeStore.HeaderIdx {
         const entry_start = self.pos;
         if (self.peek() != .LowerIdent) {
             self.store.clearScratchRecordFieldsFrom(fields_scratch_top);
-            return self.pushMalformed(AST.NodeStore.HeaderIdx, .expected_package_or_platform_name, start);
+            return self.pushMalformed(NodeStore.HeaderIdx, .expected_package_or_platform_name, start);
         }
         const name_tok = self.pos;
         self.advance();
         if (self.peek() != .OpColon) {
             self.store.clearScratchRecordFieldsFrom(fields_scratch_top);
-            return self.pushMalformed(AST.NodeStore.HeaderIdx, .expected_package_or_platform_colon, start);
+            return self.pushMalformed(NodeStore.HeaderIdx, .expected_package_or_platform_colon, start);
         }
         self.advance();
         if (self.peek() == .KwPlatform) {
             if (platform != null) {
                 self.store.clearScratchRecordFieldsFrom(fields_scratch_top);
-                return self.pushMalformed(AST.NodeStore.HeaderIdx, .multiple_platforms, start);
+                return self.pushMalformed(NodeStore.HeaderIdx, .multiple_platforms, start);
             }
             self.advance();
             if (self.peek() != .StringStart) {
                 self.store.clearScratchRecordFieldsFrom(fields_scratch_top);
-                return self.pushMalformed(AST.NodeStore.HeaderIdx, .expected_platform_string, start);
+                return self.pushMalformed(NodeStore.HeaderIdx, .expected_platform_string, start);
             }
             const value = self.parseStringExpr();
             const pidx = self.store.addRecordField(.{
@@ -714,7 +715,7 @@ pub fn parseAppHeader(self: *Parser) AST.NodeStore.HeaderIdx {
         } else {
             if (self.peek() != .StringStart) {
                 self.store.clearScratchRecordFieldsFrom(fields_scratch_top);
-                return self.pushMalformed(AST.NodeStore.HeaderIdx, .expected_package_or_platform_string, start);
+                return self.pushMalformed(NodeStore.HeaderIdx, .expected_package_or_platform_string, start);
             }
             const value = self.parseStringExpr();
             self.store.addScratchRecordField(self.store.addRecordField(.{
@@ -736,7 +737,7 @@ pub fn parseAppHeader(self: *Parser) AST.NodeStore.HeaderIdx {
     if (self.peek() != .CloseCurly) {
         self.store.clearScratchRecordFieldsFrom(fields_scratch_top);
         std.debug.print("Tokens: {any}\n", .{self.tok_buf.tokens.items(.tag)[start..self.pos]});
-        return self.pushMalformed(AST.NodeStore.HeaderIdx, .expected_package_platform_close_curly, start);
+        return self.pushMalformed(NodeStore.HeaderIdx, .expected_package_platform_close_curly, start);
     }
     self.advanceOne(); // Advance past CloseCurly
     const packages_span = self.store.recordFieldSpanFrom(fields_scratch_top);
@@ -750,7 +751,7 @@ pub fn parseAppHeader(self: *Parser) AST.NodeStore.HeaderIdx {
     self.advance();
 
     if (platform) |pidx| {
-        const header = AST.NodeStore.Header{
+        const header = NodeStore.Header{
             .app = .{
                 .platform_idx = pidx,
                 .provides = provides,
@@ -761,11 +762,11 @@ pub fn parseAppHeader(self: *Parser) AST.NodeStore.HeaderIdx {
         const idx = self.store.addHeader(header);
         return idx;
     }
-    return self.pushMalformed(AST.NodeStore.HeaderIdx, .no_platform, start);
+    return self.pushMalformed(NodeStore.HeaderIdx, .no_platform, start);
 }
 
 /// Parses an ExposedItem, adding it to the NodeStore and returning the Idx
-pub fn parseExposedItem(self: *Parser) AST.NodeStore.ExposedItemIdx {
+pub fn parseExposedItem(self: *Parser) NodeStore.ExposedItemIdx {
     const start = self.pos;
     var end = start;
     switch (self.peek()) {
@@ -776,7 +777,7 @@ pub fn parseExposedItem(self: *Parser) AST.NodeStore.ExposedItemIdx {
                 self.advance(); // Advance past KwAs
                 as = self.pos;
                 self.expect(.LowerIdent) catch {
-                    return self.pushMalformed(AST.NodeStore.ExposedItemIdx, .expected_lower_name_after_exposed_item_as, start);
+                    return self.pushMalformed(NodeStore.ExposedItemIdx, .expected_lower_name_after_exposed_item_as, start);
                 };
                 end = self.pos;
             } else {
@@ -797,7 +798,7 @@ pub fn parseExposedItem(self: *Parser) AST.NodeStore.ExposedItemIdx {
                 self.advance(); // Advance past KwAs
                 as = self.pos;
                 self.expect(.UpperIdent) catch {
-                    return self.pushMalformed(AST.NodeStore.ExposedItemIdx, .expected_upper_name_after_exposed_item_as, start);
+                    return self.pushMalformed(NodeStore.ExposedItemIdx, .expected_upper_name_after_exposed_item_as, start);
                 };
                 end = self.pos;
             } else if (self.peekNext() == .DotStar) {
@@ -819,7 +820,7 @@ pub fn parseExposedItem(self: *Parser) AST.NodeStore.ExposedItemIdx {
             return ei;
         },
         else => {
-            return self.pushMalformed(AST.NodeStore.ExposedItemIdx, .exposed_item_unexpected_token, start);
+            return self.pushMalformed(NodeStore.ExposedItemIdx, .exposed_item_unexpected_token, start);
         },
     }
 }
@@ -829,25 +830,25 @@ const StatementType = enum { top_level, in_body };
 /// Parse a top level roc statement
 ///
 /// e.g. `import Foo`
-pub fn parseTopLevelStatement(self: *Parser) ?AST.NodeStore.StatementIdx {
+pub fn parseTopLevelStatement(self: *Parser) ?NodeStore.StatementIdx {
     return self.parseStmtByType(.top_level);
 }
 
 /// parse a in-body roc statement
 ///
 /// e.g. `foo = 2 + x`
-pub fn parseStmt(self: *Parser) ?AST.NodeStore.StatementIdx {
+pub fn parseStmt(self: *Parser) ?NodeStore.StatementIdx {
     return self.parseStmtByType(.in_body);
 }
 
 /// parse a roc statement
 ///
 /// e.g. `import Foo`, or `foo = 2 + x`
-fn parseStmtByType(self: *Parser, statementType: StatementType) ?AST.NodeStore.StatementIdx {
+fn parseStmtByType(self: *Parser, statementType: StatementType) ?NodeStore.StatementIdx {
     switch (self.peek()) {
         .KwImport => {
             if (statementType != .top_level) {
-                return self.pushMalformed(AST.NodeStore.StatementIdx, .import_must_be_top_level, self.pos);
+                return self.pushMalformed(NodeStore.StatementIdx, .import_must_be_top_level, self.pos);
             }
             const start = self.pos;
             self.advance(); // Advance past KwImport
@@ -858,7 +859,7 @@ fn parseStmtByType(self: *Parser, statementType: StatementType) ?AST.NodeStore.S
                 self.advance(); // Advance past LowerIdent
             }
             if (self.peek() == .UpperIdent or (qualifier != null and (self.peek() == .NoSpaceDotUpperIdent or self.peek() == .DotUpperIdent))) {
-                var exposes: AST.NodeStore.ExposedItemSpan = .{ .span = .{ .start = 0, .len = 0 } };
+                var exposes: NodeStore.ExposedItemSpan = .{ .span = .{ .start = 0, .len = 0 } };
                 const module_name_tok = self.pos;
                 var end = self.pos;
                 if (self.peekNext() == .KwAs) {
@@ -867,7 +868,7 @@ fn parseStmtByType(self: *Parser, statementType: StatementType) ?AST.NodeStore.S
                     alias_tok = self.pos;
                     end = self.pos;
                     self.expect(.UpperIdent) catch {
-                        const malformed = self.pushMalformed(AST.NodeStore.StatementIdx, .expected_upper_name_after_import_as, start);
+                        const malformed = self.pushMalformed(NodeStore.StatementIdx, .expected_upper_name_after_import_as, start);
                         self.advance();
                         return malformed;
                     };
@@ -875,16 +876,16 @@ fn parseStmtByType(self: *Parser, statementType: StatementType) ?AST.NodeStore.S
                     self.advance(); // Advance past ident
                     self.advance(); // Advance past KwExposing
                     self.expect(.OpenSquare) catch {
-                        return self.pushMalformed(AST.NodeStore.StatementIdx, .import_exposing_no_open, start);
+                        return self.pushMalformed(NodeStore.StatementIdx, .import_exposing_no_open, start);
                     };
                     const scratch_top = self.store.scratchExposedItemTop();
-                    end = self.parseCollectionSpan(AST.NodeStore.ExposedItemIdx, .CloseSquare, AST.NodeStore.addScratchExposedItem, Parser.parseExposedItem) catch {
+                    end = self.parseCollectionSpan(NodeStore.ExposedItemIdx, .CloseSquare, NodeStore.addScratchExposedItem, Parser.parseExposedItem) catch {
                         while (self.peek() != .CloseSquare and self.peek() != .EndOfFile) {
                             self.advance();
                         }
                         self.expect(.CloseSquare) catch {};
                         self.store.clearScratchExposedItemsFrom(scratch_top);
-                        return self.pushMalformed(AST.NodeStore.StatementIdx, .import_exposing_no_close, start);
+                        return self.pushMalformed(NodeStore.StatementIdx, .import_exposing_no_close, start);
                     };
                     exposes = self.store.exposedItemSpanFrom(scratch_top);
                 } else {
@@ -925,7 +926,7 @@ fn parseStmtByType(self: *Parser, statementType: StatementType) ?AST.NodeStore.S
             self.advance();
             const patt = self.parsePattern(.alternatives_forbidden);
             self.expect(.KwIn) catch {
-                return self.pushMalformed(AST.NodeStore.StatementIdx, .for_expected_in, self.pos);
+                return self.pushMalformed(NodeStore.StatementIdx, .for_expected_in, self.pos);
             };
             const expr = self.parseExpr();
             const body = self.parseExpr();
@@ -971,16 +972,16 @@ fn parseStmtByType(self: *Parser, statementType: StatementType) ?AST.NodeStore.S
         .KwVar => {
             const start = self.pos;
             if (statementType != .in_body) {
-                return self.pushMalformed(AST.NodeStore.StatementIdx, .var_only_allowed_in_a_body, self.pos);
+                return self.pushMalformed(NodeStore.StatementIdx, .var_only_allowed_in_a_body, self.pos);
             }
             self.advance();
             if (self.peek() != .LowerIdent) {
-                return self.pushMalformed(AST.NodeStore.StatementIdx, .var_must_have_ident, self.pos);
+                return self.pushMalformed(NodeStore.StatementIdx, .var_must_have_ident, self.pos);
             }
             const name = self.pos;
             self.advance();
             self.expect(.OpAssign) catch {
-                return self.pushMalformed(AST.NodeStore.StatementIdx, .var_expected_equals, self.pos);
+                return self.pushMalformed(NodeStore.StatementIdx, .var_expected_equals, self.pos);
             };
             const body = self.parseExpr();
             const statement_idx = self.store.addStatement(.{ .@"var" = .{
@@ -1034,7 +1035,7 @@ fn parseStmtByType(self: *Parser, statementType: StatementType) ?AST.NodeStore.S
             if (statementType == .top_level) {
                 const header = self.parseTypeHeader();
                 if (self.peek() != .OpColon) {
-                    return self.pushMalformed(AST.NodeStore.StatementIdx, .expected_colon_after_type_annotation, start);
+                    return self.pushMalformed(NodeStore.StatementIdx, .expected_colon_after_type_annotation, start);
                 }
                 self.advance();
                 const anno = self.parseTypeAnno(.not_looking_for_args);
@@ -1063,7 +1064,7 @@ fn parseStmtByType(self: *Parser, statementType: StatementType) ?AST.NodeStore.S
     return statement_idx;
 }
 
-fn parseWhereConstraint(self: *Parser) ?AST.NodeStore.CollectionIdx {
+fn parseWhereConstraint(self: *Parser) ?NodeStore.CollectionIdx {
     self.expect(.KwWhere) catch {
         return null;
     };
@@ -1098,13 +1099,13 @@ const Alternatives = enum {
 };
 
 /// todo -- what does this do?
-pub fn parsePattern(self: *Parser, alternatives: Alternatives) AST.NodeStore.PatternIdx {
+pub fn parsePattern(self: *Parser, alternatives: Alternatives) NodeStore.PatternIdx {
     const outer_start = self.pos;
     const patterns_scratch_top = self.store.scratchPatternTop();
     errdefer self.store.clearScratchPatternsFrom(patterns_scratch_top);
     while (self.peek() != .EndOfFile) {
         const start = self.pos;
-        var pattern: ?AST.NodeStore.PatternIdx = null;
+        var pattern: ?NodeStore.PatternIdx = null;
         switch (self.peek()) {
             .LowerIdent => {
                 pattern = self.store.addPattern(.{ .ident = .{
@@ -1129,12 +1130,12 @@ pub fn parsePattern(self: *Parser, alternatives: Alternatives) AST.NodeStore.Pat
                     self.advance(); // Advance past NoSpaceOpenRound
                     // Parse args
                     const scratch_top = self.store.scratchPatternTop();
-                    const args_end = self.parseCollectionSpan(AST.NodeStore.PatternIdx, .CloseRound, AST.NodeStore.addScratchPattern, parsePatternWithAlts) catch {
+                    const args_end = self.parseCollectionSpan(NodeStore.PatternIdx, .CloseRound, NodeStore.addScratchPattern, parsePatternWithAlts) catch {
                         while (self.peek() != .CloseRound and self.peek() != .EndOfFile) {
                             self.advance();
                         }
                         self.store.clearScratchPatternsFrom(scratch_top);
-                        return self.pushMalformed(AST.NodeStore.PatternIdx, .pattern_unexpected_token, start);
+                        return self.pushMalformed(NodeStore.PatternIdx, .pattern_unexpected_token, start);
                     };
                     const args = self.store.patternSpanFrom(scratch_top);
                     pattern = self.store.addPattern(.{ .tag = .{
@@ -1167,12 +1168,12 @@ pub fn parsePattern(self: *Parser, alternatives: Alternatives) AST.NodeStore.Pat
                 // List
                 self.advance();
                 const scratch_top = self.store.scratchPatternTop();
-                const end = self.parseCollectionSpan(AST.NodeStore.PatternIdx, .CloseSquare, AST.NodeStore.addScratchPattern, parsePatternWithAlts) catch {
+                const end = self.parseCollectionSpan(NodeStore.PatternIdx, .CloseSquare, NodeStore.addScratchPattern, parsePatternWithAlts) catch {
                     while (self.peek() != .CloseSquare and self.peek() != .EndOfFile) {
                         self.advance();
                     }
                     self.store.clearScratchPatternsFrom(scratch_top);
-                    return self.pushMalformed(AST.NodeStore.PatternIdx, .pattern_unexpected_token, start);
+                    return self.pushMalformed(NodeStore.PatternIdx, .pattern_unexpected_token, start);
                 };
                 const patterns = self.store.patternSpanFrom(scratch_top);
 
@@ -1193,7 +1194,7 @@ pub fn parsePattern(self: *Parser, alternatives: Alternatives) AST.NodeStore.Pat
                 }
                 const fields = self.store.patternRecordFieldSpanFrom(scratch_top);
                 if (self.peek() != .CloseCurly) {
-                    return self.pushMalformed(AST.NodeStore.PatternIdx, .pattern_unexpected_token, start);
+                    return self.pushMalformed(NodeStore.PatternIdx, .pattern_unexpected_token, start);
                 }
                 const end = self.pos;
                 self.advance();
@@ -1209,7 +1210,7 @@ pub fn parsePattern(self: *Parser, alternatives: Alternatives) AST.NodeStore.Pat
                 if (self.peek() == .KwAs) {
                     self.advance();
                     if (self.peek() != .LowerIdent) {
-                        return self.pushMalformed(AST.NodeStore.PatternIdx, .pattern_unexpected_token, start);
+                        return self.pushMalformed(NodeStore.PatternIdx, .pattern_unexpected_token, start);
                     }
                     name = self.pos;
                     end = self.pos;
@@ -1229,12 +1230,12 @@ pub fn parsePattern(self: *Parser, alternatives: Alternatives) AST.NodeStore.Pat
             .OpenRound, .NoSpaceOpenRound => {
                 self.advance();
                 const scratch_top = self.store.scratchPatternTop();
-                const end = self.parseCollectionSpan(AST.NodeStore.PatternIdx, .CloseRound, AST.NodeStore.addScratchPattern, parsePatternWithAlts) catch {
+                const end = self.parseCollectionSpan(NodeStore.PatternIdx, .CloseRound, NodeStore.addScratchPattern, parsePatternWithAlts) catch {
                     while (self.peek() != .CloseRound and self.peek() != .EndOfFile) {
                         self.advance();
                     }
                     self.store.clearScratchPatternsFrom(scratch_top);
-                    return self.pushMalformed(AST.NodeStore.PatternIdx, .pattern_unexpected_token, start);
+                    return self.pushMalformed(NodeStore.PatternIdx, .pattern_unexpected_token, start);
                 };
                 const patterns = self.store.patternSpanFrom(scratch_top);
 
@@ -1244,7 +1245,7 @@ pub fn parsePattern(self: *Parser, alternatives: Alternatives) AST.NodeStore.Pat
                 } });
             },
             else => {
-                return self.pushMalformed(AST.NodeStore.PatternIdx, .pattern_unexpected_token, self.pos);
+                return self.pushMalformed(NodeStore.PatternIdx, .pattern_unexpected_token, self.pos);
             },
         }
 
@@ -1264,7 +1265,7 @@ pub fn parsePattern(self: *Parser, alternatives: Alternatives) AST.NodeStore.Pat
         }
     }
     if ((self.store.scratchPatternTop() - patterns_scratch_top) == 0) {
-        return self.store.addMalformed(AST.NodeStore.PatternIdx, .pattern_unexpected_eof, .{ .start = outer_start, .end = self.pos });
+        return self.store.addMalformed(NodeStore.PatternIdx, .pattern_unexpected_eof, .{ .start = outer_start, .end = self.pos });
     }
     const last_pattern = self.store.scratch_patterns.items.items[self.store.scratchPatternTop() - 1];
     const last_pattern_region = self.store.nodes.items.items(.region)[last_pattern.id];
@@ -1275,15 +1276,15 @@ pub fn parsePattern(self: *Parser, alternatives: Alternatives) AST.NodeStore.Pat
     } });
 }
 
-fn parsePatternNoAlts(self: *Parser) AST.NodeStore.PatternIdx {
+fn parsePatternNoAlts(self: *Parser) NodeStore.PatternIdx {
     return self.parsePattern(.alternatives_forbidden);
 }
-fn parsePatternWithAlts(self: *Parser) AST.NodeStore.PatternIdx {
+fn parsePatternWithAlts(self: *Parser) NodeStore.PatternIdx {
     return self.parsePattern(.alternatives_allowed);
 }
 
 /// todo
-pub fn parsePatternRecordField(self: *Parser, alternatives: Alternatives) AST.NodeStore.PatternRecordFieldIdx {
+pub fn parsePatternRecordField(self: *Parser, alternatives: Alternatives) NodeStore.PatternRecordFieldIdx {
     const field_start = self.pos;
     if (self.peek() == .DoubleDot) {
         self.advance();
@@ -1303,16 +1304,16 @@ pub fn parsePatternRecordField(self: *Parser, alternatives: Alternatives) AST.No
         while (self.peek() != .CloseCurly and self.peek() != .EndOfFile) {
             self.advance();
         }
-        return self.pushMalformed(AST.NodeStore.PatternRecordFieldIdx, .expected_lower_ident_pat_field_name, field_start);
+        return self.pushMalformed(NodeStore.PatternRecordFieldIdx, .expected_lower_ident_pat_field_name, field_start);
     }
     const name = self.pos;
     self.advance();
-    var value: ?AST.NodeStore.PatternIdx = null;
+    var value: ?NodeStore.PatternIdx = null;
     if (self.peek() != .OpColon and (self.peekNext() != .Comma or self.peekNext() != .CloseCurly)) {
         while (self.peek() != .CloseCurly and self.peek() != .EndOfFile) {
             self.advance();
         }
-        return self.pushMalformed(AST.NodeStore.PatternRecordFieldIdx, .expected_colon_after_pat_field_name, field_start);
+        return self.pushMalformed(NodeStore.PatternRecordFieldIdx, .expected_colon_after_pat_field_name, field_start);
     }
     self.advance();
     if (self.peekNext() != .Comma or self.peekNext() != .CloseCurly) {
@@ -1329,14 +1330,14 @@ pub fn parsePatternRecordField(self: *Parser, alternatives: Alternatives) AST.No
 }
 
 /// todo
-pub fn parseExpr(self: *Parser) AST.NodeStore.ExprIdx {
+pub fn parseExpr(self: *Parser) NodeStore.ExprIdx {
     return self.parseExprWithBp(0);
 }
 
 /// todo
-pub fn parseExprWithBp(self: *Parser, min_bp: u8) AST.NodeStore.ExprIdx {
+pub fn parseExprWithBp(self: *Parser, min_bp: u8) NodeStore.ExprIdx {
     const start = self.pos;
-    var expr: ?AST.NodeStore.ExprIdx = null;
+    var expr: ?NodeStore.ExprIdx = null;
     const token = self.peek();
     switch (token) {
         .UpperIdent => {
@@ -1393,12 +1394,12 @@ pub fn parseExprWithBp(self: *Parser, min_bp: u8) AST.NodeStore.ExprIdx {
         .OpenSquare => {
             self.advance();
             const scratch_top = self.store.scratchExprTop();
-            const list_end = self.parseCollectionSpan(AST.NodeStore.ExprIdx, .CloseSquare, AST.NodeStore.addScratchExpr, parseExpr) catch {
+            const list_end = self.parseCollectionSpan(NodeStore.ExprIdx, .CloseSquare, NodeStore.addScratchExpr, parseExpr) catch {
                 while (self.peek() != .CloseSquare and self.peek() != .EndOfFile) {
                     self.advance();
                 }
                 self.store.clearScratchExprsFrom(scratch_top);
-                return self.pushMalformed(AST.NodeStore.ExprIdx, .expected_expr_close_square_or_comma, self.pos);
+                return self.pushMalformed(NodeStore.ExprIdx, .expected_expr_close_square_or_comma, self.pos);
             };
             const items = self.store.exprSpanFrom(scratch_top);
             expr = self.store.addExpr(.{ .list = .{
@@ -1410,12 +1411,12 @@ pub fn parseExprWithBp(self: *Parser, min_bp: u8) AST.NodeStore.ExprIdx {
             self.advance();
             // TODO: Parenthesized expressions
             const scratch_top = self.store.scratchExprTop();
-            const end = self.parseCollectionSpan(AST.NodeStore.ExprIdx, .CloseRound, AST.NodeStore.addScratchExpr, parseExpr) catch {
+            const end = self.parseCollectionSpan(NodeStore.ExprIdx, .CloseRound, NodeStore.addScratchExpr, parseExpr) catch {
                 while (self.peek() != .CloseRound and self.peek() != .EndOfFile) {
                     self.advance();
                 }
                 self.store.clearScratchExprsFrom(scratch_top);
-                return self.pushMalformed(AST.NodeStore.ExprIdx, .expected_expr_close_round_or_comma, self.pos);
+                return self.pushMalformed(NodeStore.ExprIdx, .expected_expr_close_round_or_comma, self.pos);
             };
             const items = self.store.exprSpanFrom(scratch_top);
             expr = self.store.addExpr(.{ .tuple = .{
@@ -1430,9 +1431,9 @@ pub fn parseExprWithBp(self: *Parser, min_bp: u8) AST.NodeStore.ExprIdx {
                 // This is the best guesstimation of this being a Record for now.  I believe we have to have a NoSpaceOpColon
                 // for this to be full-proof without backtracking.
                 const scratch_top = self.store.scratchRecordFieldTop();
-                const end = self.parseCollectionSpan(AST.NodeStore.RecordFieldIdx, .CloseCurly, AST.NodeStore.addScratchRecordField, parseRecordField) catch {
+                const end = self.parseCollectionSpan(NodeStore.RecordFieldIdx, .CloseCurly, NodeStore.addScratchRecordField, parseRecordField) catch {
                     self.store.clearScratchRecordFieldsFrom(scratch_top);
-                    return self.pushMalformed(AST.NodeStore.ExprIdx, .expected_expr_close_curly_or_comma, self.pos);
+                    return self.pushMalformed(NodeStore.ExprIdx, .expected_expr_close_curly_or_comma, self.pos);
                 };
                 const fields = self.store.recordFieldSpanFrom(scratch_top);
                 expr = self.store.addExpr(.{ .record = .{
@@ -1464,9 +1465,9 @@ pub fn parseExprWithBp(self: *Parser, min_bp: u8) AST.NodeStore.ExprIdx {
         .OpBar => {
             self.advance();
             const scratch_top = self.store.scratchPatternTop();
-            _ = self.parseCollectionSpan(AST.NodeStore.PatternIdx, .OpBar, AST.NodeStore.addScratchPattern, parsePatternNoAlts) catch {
+            _ = self.parseCollectionSpan(NodeStore.PatternIdx, .OpBar, NodeStore.addScratchPattern, parsePatternNoAlts) catch {
                 self.store.clearScratchPatternsFrom(scratch_top);
-                return self.pushMalformed(AST.NodeStore.ExprIdx, .expected_expr_bar, self.pos);
+                return self.pushMalformed(NodeStore.ExprIdx, .expected_expr_bar, self.pos);
             };
             const args = self.store.patternSpanFrom(scratch_top);
 
@@ -1483,7 +1484,7 @@ pub fn parseExprWithBp(self: *Parser, min_bp: u8) AST.NodeStore.ExprIdx {
             const condition = self.parseExpr();
             const then = self.parseExpr();
             if (self.peek() != .KwElse) {
-                return self.pushMalformed(AST.NodeStore.ExprIdx, .no_else, self.pos);
+                return self.pushMalformed(NodeStore.ExprIdx, .no_else, self.pos);
             }
             self.advance();
             const else_idx = self.parseExpr();
@@ -1499,7 +1500,7 @@ pub fn parseExprWithBp(self: *Parser, min_bp: u8) AST.NodeStore.ExprIdx {
             const e = self.parseExpr();
 
             self.expect(.OpenCurly) catch {
-                return self.pushMalformed(AST.NodeStore.ExprIdx, .expected_open_curly_after_match, self.pos);
+                return self.pushMalformed(NodeStore.ExprIdx, .expected_open_curly_after_match, self.pos);
             };
             const scratch_top = self.store.scratchWhenBranchTop();
             while (self.peek() != .CloseCurly and self.peek() != .EndOfFile) {
@@ -1510,7 +1511,7 @@ pub fn parseExprWithBp(self: *Parser, min_bp: u8) AST.NodeStore.ExprIdx {
             }
             const branches = self.store.whenBranchSpanFrom(scratch_top);
             if (self.peek() != .CloseCurly) {
-                return self.pushMalformed(AST.NodeStore.ExprIdx, .expected_close_curly_at_end_of_match, self.pos);
+                return self.pushMalformed(NodeStore.ExprIdx, .expected_close_curly_at_end_of_match, self.pos);
             }
             self.advance();
             expr = self.store.addExpr(.{ .match = .{
@@ -1534,7 +1535,7 @@ pub fn parseExprWithBp(self: *Parser, min_bp: u8) AST.NodeStore.ExprIdx {
             self.advance();
         },
         else => {
-            return self.pushMalformed(AST.NodeStore.ExprIdx, .expr_unexpected_token, start);
+            return self.pushMalformed(NodeStore.ExprIdx, .expr_unexpected_token, start);
         },
     }
     if (expr) |e| {
@@ -1542,7 +1543,7 @@ pub fn parseExprWithBp(self: *Parser, min_bp: u8) AST.NodeStore.ExprIdx {
         while (self.peek() == .NoSpaceDotInt or self.peek() == .NoSpaceDotLowerIdent or self.peek() == .DotLowerIdent or self.peek() == .OpArrow) {
             const tok = self.peek();
             if (tok == .NoSpaceDotInt) {
-                return self.pushMalformed(AST.NodeStore.ExprIdx, .expr_no_space_dot_int, self.pos);
+                return self.pushMalformed(NodeStore.ExprIdx, .expr_no_space_dot_int, self.pos);
             } else if (self.peek() == .OpArrow) {
                 const s = self.pos;
                 self.advance();
@@ -1574,7 +1575,7 @@ pub fn parseExprWithBp(self: *Parser, min_bp: u8) AST.NodeStore.ExprIdx {
                         .right = ident_suffixed,
                     } });
                 } else {
-                    return self.pushMalformed(AST.NodeStore.ExprIdx, .expr_arrow_expects_ident, self.pos);
+                    return self.pushMalformed(NodeStore.ExprIdx, .expr_arrow_expects_ident, self.pos);
                 }
             } else { // NoSpaceDotLowerIdent
                 const s = self.pos;
@@ -1612,19 +1613,19 @@ pub fn parseExprWithBp(self: *Parser, min_bp: u8) AST.NodeStore.ExprIdx {
         }
         return expression;
     }
-    return self.store.addMalformed(AST.NodeStore.ExprIdx, .expr_unexpected_token, .{ .start = start, .end = self.pos });
+    return self.store.addMalformed(NodeStore.ExprIdx, .expr_unexpected_token, .{ .start = start, .end = self.pos });
 }
 
 /// todo
-fn parseExprSuffix(self: *Parser, start: u32, e: AST.NodeStore.ExprIdx) AST.NodeStore.ExprIdx {
+fn parseExprSuffix(self: *Parser, start: u32, e: NodeStore.ExprIdx) NodeStore.ExprIdx {
     var expression = e;
     // Check for an apply...
     if (self.peek() == .NoSpaceOpenRound) {
         self.advance();
         const scratch_top = self.store.scratchExprTop();
-        const end = self.parseCollectionSpan(AST.NodeStore.ExprIdx, .CloseRound, AST.NodeStore.addScratchExpr, parseExpr) catch {
+        const end = self.parseCollectionSpan(NodeStore.ExprIdx, .CloseRound, NodeStore.addScratchExpr, parseExpr) catch {
             self.store.clearScratchExprsFrom(scratch_top);
-            return self.pushMalformed(AST.NodeStore.ExprIdx, .expected_expr_apply_close_round, start);
+            return self.pushMalformed(NodeStore.ExprIdx, .expected_expr_apply_close_round, start);
         };
         const args = self.store.exprSpanFrom(scratch_top);
 
@@ -1646,13 +1647,13 @@ fn parseExprSuffix(self: *Parser, start: u32, e: AST.NodeStore.ExprIdx) AST.Node
 }
 
 /// todo
-pub fn parseRecordField(self: *Parser) AST.NodeStore.RecordFieldIdx {
+pub fn parseRecordField(self: *Parser) NodeStore.RecordFieldIdx {
     const start = self.pos;
     self.expect(.LowerIdent) catch {
-        return self.pushMalformed(AST.NodeStore.RecordFieldIdx, .expected_expr_record_field_name, start);
+        return self.pushMalformed(NodeStore.RecordFieldIdx, .expected_expr_record_field_name, start);
     };
     const name = start;
-    var value: ?AST.NodeStore.ExprIdx = null;
+    var value: ?NodeStore.ExprIdx = null;
     if (self.peek() == .OpColon) {
         self.advance();
         value = self.parseExpr();
@@ -1667,7 +1668,7 @@ pub fn parseRecordField(self: *Parser) AST.NodeStore.RecordFieldIdx {
 }
 
 /// todo
-pub fn parseBranch(self: *Parser) AST.NodeStore.WhenBranchIdx {
+pub fn parseBranch(self: *Parser) NodeStore.WhenBranchIdx {
     const start = self.pos;
     const p = self.parsePattern(.alternatives_allowed);
     if (self.peek() == .OpFatArrow) {
@@ -1682,7 +1683,7 @@ pub fn parseBranch(self: *Parser) AST.NodeStore.WhenBranchIdx {
 }
 
 /// todo
-pub fn parseStringExpr(self: *Parser) AST.NodeStore.ExprIdx {
+pub fn parseStringExpr(self: *Parser) NodeStore.ExprIdx {
     std.debug.assert(self.peek() == .StringStart);
     const start = self.pos;
     // Start parsing string with possible interpolations
@@ -1714,13 +1715,13 @@ pub fn parseStringExpr(self: *Parser) AST.NodeStore.ExprIdx {
                     self.advanceOne(); // Advance ONLY past the Newline
                 }
                 if (self.peek() != .CloseStringInterpolation) {
-                    return self.pushMalformed(AST.NodeStore.ExprIdx, .string_expected_close_interpolation, start);
+                    return self.pushMalformed(NodeStore.ExprIdx, .string_expected_close_interpolation, start);
                 }
                 self.advanceOne(); // Advance ONLY past the CloseString Interpolation
             },
             else => {
                 // Something is broken in the tokenizer if we get here!
-                return self.pushMalformed(AST.NodeStore.ExprIdx, .string_unexpected_token, self.pos);
+                return self.pushMalformed(NodeStore.ExprIdx, .string_unexpected_token, self.pos);
             },
         }
     }
@@ -1739,7 +1740,7 @@ pub fn parseStringExpr(self: *Parser) AST.NodeStore.ExprIdx {
 }
 
 /// todo
-pub fn parseStringPattern(self: *Parser) AST.NodeStore.PatternIdx {
+pub fn parseStringPattern(self: *Parser) NodeStore.PatternIdx {
     const start = self.pos;
     const inner = parseStringExpr(self);
     const patt_idx = self.store.addPattern(.{ .string = .{
@@ -1751,7 +1752,7 @@ pub fn parseStringPattern(self: *Parser) AST.NodeStore.PatternIdx {
 }
 
 /// todo
-pub fn parseTypeHeader(self: *Parser) AST.NodeStore.TypeHeaderIdx {
+pub fn parseTypeHeader(self: *Parser) NodeStore.TypeHeaderIdx {
     const start = self.pos;
     std.debug.assert(self.peek() == .UpperIdent);
     self.advance(); // Advance past UpperIdent
@@ -1767,9 +1768,9 @@ pub fn parseTypeHeader(self: *Parser) AST.NodeStore.TypeHeaderIdx {
     }
     self.advance();
     const scratch_top = self.store.scratchTypeAnnoTop();
-    const end = self.parseCollectionSpan(AST.NodeStore.TypeAnnoIdx, .CloseRound, AST.NodeStore.addScratchTypeAnno, Parser.parseTypeIdent) catch {
+    const end = self.parseCollectionSpan(NodeStore.TypeAnnoIdx, .CloseRound, NodeStore.addScratchTypeAnno, Parser.parseTypeIdent) catch {
         self.store.clearScratchTypeAnnosFrom(scratch_top);
-        return self.pushMalformed(AST.NodeStore.TypeHeaderIdx, .expected_ty_anno_end, start);
+        return self.pushMalformed(NodeStore.TypeHeaderIdx, .expected_ty_anno_end, start);
     };
     const args = self.store.typeAnnoSpanFrom(scratch_top);
     return self.store.addTypeHeader(.{
@@ -1779,7 +1780,7 @@ pub fn parseTypeHeader(self: *Parser) AST.NodeStore.TypeHeaderIdx {
     });
 }
 
-fn parseTypeIdent(self: *Parser) AST.NodeStore.TypeAnnoIdx {
+fn parseTypeIdent(self: *Parser) NodeStore.TypeAnnoIdx {
     if (self.peek() == .LowerIdent) {
         const tok = self.pos;
         self.advance();
@@ -1788,7 +1789,7 @@ fn parseTypeIdent(self: *Parser) AST.NodeStore.TypeAnnoIdx {
             .tok = tok,
         } });
     }
-    return self.pushMalformed(AST.NodeStore.TypeAnnoIdx, .invalid_type_arg, self.pos);
+    return self.pushMalformed(NodeStore.TypeAnnoIdx, .invalid_type_arg, self.pos);
 }
 
 const TyFnArgs = enum {
@@ -1797,9 +1798,9 @@ const TyFnArgs = enum {
 };
 
 /// Parse a type annotation, e.g. `Foo(a) : (a,Str,I64)`
-pub fn parseTypeAnno(self: *Parser, looking_for_args: TyFnArgs) AST.NodeStore.TypeAnnoIdx {
+pub fn parseTypeAnno(self: *Parser, looking_for_args: TyFnArgs) NodeStore.TypeAnnoIdx {
     const start = self.pos;
-    var anno: ?AST.NodeStore.TypeAnnoIdx = null;
+    var anno: ?NodeStore.TypeAnnoIdx = null;
 
     switch (self.peek()) {
         .UpperIdent => {
@@ -1823,11 +1824,11 @@ pub fn parseTypeAnno(self: *Parser, looking_for_args: TyFnArgs) AST.NodeStore.Ty
                 self.advance(); // Advance past NoSpaceOpenRound
                 const scratch_top = self.store.scratchTypeAnnoTop();
                 self.store.addScratchTypeAnno(anno orelse {
-                    return self.store.addMalformed(AST.NodeStore.TypeAnnoIdx, .ty_anno_unexpected_token, .{ .start = start, .end = self.pos });
+                    return self.store.addMalformed(NodeStore.TypeAnnoIdx, .ty_anno_unexpected_token, .{ .start = start, .end = self.pos });
                 });
-                const end = self.parseCollectionSpan(AST.NodeStore.TypeAnnoIdx, .CloseRound, AST.NodeStore.addScratchTypeAnno, parseTypeAnnoInCollection) catch {
+                const end = self.parseCollectionSpan(NodeStore.TypeAnnoIdx, .CloseRound, NodeStore.addScratchTypeAnno, parseTypeAnnoInCollection) catch {
                     self.store.clearScratchTypeAnnosFrom(scratch_top);
-                    return self.pushMalformed(AST.NodeStore.TypeAnnoIdx, .expected_ty_apply_close_round, start);
+                    return self.pushMalformed(NodeStore.TypeAnnoIdx, .expected_ty_apply_close_round, start);
                 };
 
                 anno = self.store.addTypeAnno(.{ .apply = .{
@@ -1866,7 +1867,7 @@ pub fn parseTypeAnno(self: *Parser, looking_for_args: TyFnArgs) AST.NodeStore.Ty
                 const ret_region = self.store.nodes.items.items(.region)[ret.id];
                 if (self.peek() != .CloseRound) {
                     self.store.clearScratchTypeAnnosFrom(scratch_top);
-                    return self.pushMalformed(AST.NodeStore.TypeAnnoIdx, .expected_ty_anno_end_of_function, start);
+                    return self.pushMalformed(NodeStore.TypeAnnoIdx, .expected_ty_anno_end_of_function, start);
                 }
                 const function = self.store.addTypeAnno(.{ .@"fn" = .{
                     .args = args,
@@ -1883,7 +1884,7 @@ pub fn parseTypeAnno(self: *Parser, looking_for_args: TyFnArgs) AST.NodeStore.Ty
             }
             if (self.peek() != .CloseRound) {
                 self.store.clearScratchTypeAnnosFrom(scratch_top);
-                return self.pushMalformed(AST.NodeStore.TypeAnnoIdx, .expected_ty_anno_end, start);
+                return self.pushMalformed(NodeStore.TypeAnnoIdx, .expected_ty_anno_end, start);
             }
             const end = self.pos;
             self.advance(); // Advance past CloseRound
@@ -1896,9 +1897,9 @@ pub fn parseTypeAnno(self: *Parser, looking_for_args: TyFnArgs) AST.NodeStore.Ty
         .OpenCurly => {
             self.advance(); // Advance past OpenCurly
             const scratch_top = self.store.scratchAnnoRecordFieldTop();
-            const end = self.parseCollectionSpan(AST.NodeStore.AnnoRecordFieldIdx, .CloseCurly, AST.NodeStore.addScratchAnnoRecordField, parseAnnoRecordField) catch {
+            const end = self.parseCollectionSpan(NodeStore.AnnoRecordFieldIdx, .CloseCurly, NodeStore.addScratchAnnoRecordField, parseAnnoRecordField) catch {
                 self.store.clearScratchAnnoRecordFieldsFrom(scratch_top);
-                return self.pushMalformed(AST.NodeStore.TypeAnnoIdx, .expected_ty_close_curly_or_comma, self.pos);
+                return self.pushMalformed(NodeStore.TypeAnnoIdx, .expected_ty_close_curly_or_comma, self.pos);
             };
             const fields = self.store.annoRecordFieldSpanFrom(scratch_top);
             anno = self.store.addTypeAnno(.{ .record = .{
@@ -1909,9 +1910,9 @@ pub fn parseTypeAnno(self: *Parser, looking_for_args: TyFnArgs) AST.NodeStore.Ty
         .OpenSquare => {
             self.advance(); // Advance past OpenSquare
             const scratch_top = self.store.scratchTypeAnnoTop();
-            const end = self.parseCollectionSpan(AST.NodeStore.TypeAnnoIdx, .CloseSquare, AST.NodeStore.addScratchTypeAnno, parseTypeAnnoInCollection) catch {
+            const end = self.parseCollectionSpan(NodeStore.TypeAnnoIdx, .CloseSquare, NodeStore.addScratchTypeAnno, parseTypeAnnoInCollection) catch {
                 self.store.clearScratchTypeAnnosFrom(scratch_top);
-                return self.pushMalformed(AST.NodeStore.TypeAnnoIdx, .expected_ty_close_square_or_comma, self.pos);
+                return self.pushMalformed(NodeStore.TypeAnnoIdx, .expected_ty_close_square_or_comma, self.pos);
             };
             const tags = self.store.typeAnnoSpanFrom(scratch_top);
             anno = self.store.addTypeAnno(.{ .tag_union = .{
@@ -1927,7 +1928,7 @@ pub fn parseTypeAnno(self: *Parser, looking_for_args: TyFnArgs) AST.NodeStore.Ty
             self.advance(); // Advance past Underscore
         },
         else => {
-            return self.pushMalformed(AST.NodeStore.TypeAnnoIdx, .ty_anno_unexpected_token, self.pos);
+            return self.pushMalformed(NodeStore.TypeAnnoIdx, .ty_anno_unexpected_token, self.pos);
         },
     }
 
@@ -1950,7 +1951,7 @@ pub fn parseTypeAnno(self: *Parser, looking_for_args: TyFnArgs) AST.NodeStore.Ty
             }
             const args = self.store.typeAnnoSpanFrom(scratch_top);
             if (self.peek() != .OpArrow and self.peek() != .OpFatArrow) {
-                return self.pushMalformed(AST.NodeStore.TypeAnnoIdx, .expected_arrow, start);
+                return self.pushMalformed(NodeStore.TypeAnnoIdx, .expected_arrow, start);
             }
             const effectful = self.peek() == .OpFatArrow;
             self.advance(); // Advance past arrow
@@ -1967,22 +1968,22 @@ pub fn parseTypeAnno(self: *Parser, looking_for_args: TyFnArgs) AST.NodeStore.Ty
         return an;
     }
 
-    return self.store.addMalformed(AST.NodeStore.TypeAnnoIdx, .ty_anno_unexpected_token, .{ .start = start, .end = self.pos });
+    return self.store.addMalformed(NodeStore.TypeAnnoIdx, .ty_anno_unexpected_token, .{ .start = start, .end = self.pos });
 }
 
 /// todo
-pub fn parseTypeAnnoInCollection(self: *Parser) AST.NodeStore.TypeAnnoIdx {
+pub fn parseTypeAnnoInCollection(self: *Parser) NodeStore.TypeAnnoIdx {
     return self.parseTypeAnno(.looking_for_args);
 }
 
 /// todo
-pub fn parseAnnoRecordField(self: *Parser) AST.NodeStore.AnnoRecordFieldIdx {
+pub fn parseAnnoRecordField(self: *Parser) NodeStore.AnnoRecordFieldIdx {
     const field_start = self.pos;
     if (self.peek() != .LowerIdent) {
         while (self.peek() != .CloseCurly and self.peek() != .Comma and self.peek() != .EndOfFile) {
             self.advance(); // Advance until we end this field or the record
         }
-        return self.pushMalformed(AST.NodeStore.AnnoRecordFieldIdx, .expected_type_field_name, field_start);
+        return self.pushMalformed(NodeStore.AnnoRecordFieldIdx, .expected_type_field_name, field_start);
     }
     const name = self.pos;
     self.advance(); // Advance past LowerIdent
@@ -1990,7 +1991,7 @@ pub fn parseAnnoRecordField(self: *Parser) AST.NodeStore.AnnoRecordFieldIdx {
         while (self.peek() != .CloseCurly and self.peek() != .Comma and self.peek() != .EndOfFile) {
             self.advance(); // Advance until we end this field or the record
         }
-        return self.pushMalformed(AST.NodeStore.AnnoRecordFieldIdx, .expected_colon_after_type_field_name, field_start);
+        return self.pushMalformed(NodeStore.AnnoRecordFieldIdx, .expected_colon_after_type_field_name, field_start);
     }
     self.advance(); // Advance past OpColon
     const ty = self.parseTypeAnno(.not_looking_for_args);
@@ -2007,14 +2008,14 @@ pub fn parseAnnoRecordField(self: *Parser) AST.NodeStore.AnnoRecordFieldIdx {
 /// e.g. `a.hash(hasher) -> hasher`
 /// e.g. `hasher.Hasher`
 /// e.g. `module(a).decode(List(U8)) -> a`
-pub fn parseWhereClause(self: *Parser) AST.NodeStore.WhereClauseIdx {
+pub fn parseWhereClause(self: *Parser) NodeStore.WhereClauseIdx {
     const start = self.pos;
     if (self.peek() == .KwModule) {
         // Parsing a mod_method clause
         self.advance();
         self.expect(.NoSpaceOpenRound) catch {
             return self.store.addMalformed(
-                AST.NodeStore.WhereClauseIdx,
+                NodeStore.WhereClauseIdx,
                 .where_expected_mod_open,
                 .{ .start = start, .end = self.pos },
             );
@@ -2022,14 +2023,14 @@ pub fn parseWhereClause(self: *Parser) AST.NodeStore.WhereClauseIdx {
         const var_tok = self.pos;
         self.expect(.LowerIdent) catch {
             return self.store.addMalformed(
-                AST.NodeStore.WhereClauseIdx,
+                NodeStore.WhereClauseIdx,
                 .where_expected_var,
                 .{ .start = start, .end = self.pos },
             );
         };
         self.expect(.CloseRound) catch {
             return self.store.addMalformed(
-                AST.NodeStore.WhereClauseIdx,
+                NodeStore.WhereClauseIdx,
                 .where_expected_mod_close,
                 .{ .start = start, .end = self.pos },
             );
@@ -2037,7 +2038,7 @@ pub fn parseWhereClause(self: *Parser) AST.NodeStore.WhereClauseIdx {
         const name_tok = self.pos;
         if (self.peek() != .NoSpaceDotLowerIdent and self.peek() != .DotLowerIdent) {
             return self.store.addMalformed(
-                AST.NodeStore.WhereClauseIdx,
+                NodeStore.WhereClauseIdx,
                 .where_expected_method_or_alias_name,
                 .{ .start = start, .end = self.pos },
             );
@@ -2046,21 +2047,21 @@ pub fn parseWhereClause(self: *Parser) AST.NodeStore.WhereClauseIdx {
         const arg_start = self.pos;
         self.expect(.NoSpaceOpenRound) catch {
             return self.store.addMalformed(
-                AST.NodeStore.WhereClauseIdx,
+                NodeStore.WhereClauseIdx,
                 .where_expected_arg_open,
                 .{ .start = start, .end = self.pos },
             );
         };
         const ty_anno_top = self.store.scratchTypeAnnoTop();
         const arg_end = self.parseCollectionSpan(
-            AST.NodeStore.TypeAnnoIdx,
+            NodeStore.TypeAnnoIdx,
             .CloseRound,
-            AST.NodeStore.addScratchTypeAnno,
+            NodeStore.addScratchTypeAnno,
             Parser.parseTypeAnnoInCollection,
         ) catch {
             self.store.clearScratchTypeAnnosFrom(ty_anno_top);
             return self.store.addMalformed(
-                AST.NodeStore.WhereClauseIdx,
+                NodeStore.WhereClauseIdx,
                 .where_expected_arg_close,
                 .{ .start = start, .end = self.pos },
             );
@@ -2075,7 +2076,7 @@ pub fn parseWhereClause(self: *Parser) AST.NodeStore.WhereClauseIdx {
         );
         self.expect(.OpArrow) catch {
             return self.store.addMalformed(
-                AST.NodeStore.WhereClauseIdx,
+                NodeStore.WhereClauseIdx,
                 .where_expected_method_arrow,
                 .{ .start = start, .end = self.pos },
             );
@@ -2098,21 +2099,21 @@ pub fn parseWhereClause(self: *Parser) AST.NodeStore.WhereClauseIdx {
             const arg_start = self.pos;
             self.expect(.NoSpaceOpenRound) catch {
                 return self.store.addMalformed(
-                    AST.NodeStore.WhereClauseIdx,
+                    NodeStore.WhereClauseIdx,
                     .where_expected_arg_open,
                     .{ .start = start, .end = self.pos },
                 );
             };
             const ty_anno_top = self.store.scratchTypeAnnoTop();
             const arg_end = self.parseCollectionSpan(
-                AST.NodeStore.TypeAnnoIdx,
+                NodeStore.TypeAnnoIdx,
                 .CloseRound,
-                AST.NodeStore.addScratchTypeAnno,
+                NodeStore.addScratchTypeAnno,
                 Parser.parseTypeAnnoInCollection,
             ) catch {
                 self.store.clearScratchTypeAnnosFrom(ty_anno_top);
                 return self.store.addMalformed(
-                    AST.NodeStore.WhereClauseIdx,
+                    NodeStore.WhereClauseIdx,
                     .where_expected_arg_close,
                     .{ .start = start, .end = self.pos },
                 );
@@ -2127,7 +2128,7 @@ pub fn parseWhereClause(self: *Parser) AST.NodeStore.WhereClauseIdx {
             );
             self.expect(.OpArrow) catch {
                 return self.store.addMalformed(
-                    AST.NodeStore.WhereClauseIdx,
+                    NodeStore.WhereClauseIdx,
                     .where_expected_method_arrow,
                     .{ .start = start, .end = self.pos },
                 );
@@ -2150,7 +2151,7 @@ pub fn parseWhereClause(self: *Parser) AST.NodeStore.WhereClauseIdx {
             } });
         } else {
             return self.store.addMalformed(
-                AST.NodeStore.WhereClauseIdx,
+                NodeStore.WhereClauseIdx,
                 .where_expected_method_or_alias_name,
                 .{ .start = start, .end = self.pos },
             );
@@ -2158,7 +2159,7 @@ pub fn parseWhereClause(self: *Parser) AST.NodeStore.WhereClauseIdx {
     } else {
         // malformed
         return self.store.addMalformed(
-            AST.NodeStore.WhereClauseIdx,
+            NodeStore.WhereClauseIdx,
             .where_expected_var_or_module,
             .{ .start = start, .end = self.pos },
         );
